Move AttachmentsPickerDialogStyle into MessageComposerViewStyle
