+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/composer/attachment/picker/AttachmentsPickerDialogStyle.kt
@@ -16,21 +16,10 @@
 
 package io.getstream.chat.android.ui.feature.messages.composer.attachment.picker
 
-import android.content.Context
 import android.content.res.ColorStateList
-import android.graphics.Typeface
 import android.graphics.drawable.Drawable
-import android.util.AttributeSet
 import androidx.annotation.ColorInt
-import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.font.TextStyle
-import io.getstream.chat.android.ui.helper.TransformStyle
-import io.getstream.chat.android.ui.utils.extensions.getColorCompat
-import io.getstream.chat.android.ui.utils.extensions.getColorOrNull
-import io.getstream.chat.android.ui.utils.extensions.getColorStateListCompat
-import io.getstream.chat.android.ui.utils.extensions.getDimension
-import io.getstream.chat.android.ui.utils.extensions.getDrawableCompat
-import io.getstream.chat.android.ui.utils.extensions.use
 
 /**
  * Style for [AttachmentsPickerDialogFragment].
@@ -107,316 +96,4 @@ public data class AttachmentsPickerDialogStyle(
     val cameraAttachmentsTabIconDrawable: Drawable,
     val allowAccessToCameraButtonText: String,
     val allowAccessToCameraIconDrawable: Drawable,
-) {
-    public companion object {
-        internal operator fun invoke(context: Context, attrs: AttributeSet?): AttachmentsPickerDialogStyle {
-            context.obtainStyledAttributes(
-                attrs,
-                R.styleable.AttachmentsPickerDialog,
-                R.attr.streamUiAttachmentsPickerDialogStyle,
-                R.style.StreamUi_AttachmentsPickerDialog,
-            ).use { a ->
-                val attachmentsPickerBackgroundColor = a.getColor(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerBackgroundColor,
-                    context.getColorCompat(R.color.stream_ui_white_smoke)
-                )
-
-                val allowAccessButtonTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessButtonTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_large)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessButtonTextColor,
-                        context.getColorCompat(R.color.stream_ui_accent_blue)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessButtonTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessButtonTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessButtonTextStyle,
-                        Typeface.BOLD
-                    )
-                    .build()
-
-                val submitAttachmentsButtonIconDrawable = a.getDrawableCompat(
-                    context,
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerSubmitAttachmentsButtonIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_ic_next)!!
-
-                val attachmentTabToggleButtonStateList = a.getColorStateList(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAttachmentTabToggleButtonStateList
-                ) ?: context.getColorStateListCompat(R.color.stream_ui_attachment_tab_button)
-
-                /**
-                 * Media attachments tab
-                 */
-                val mediaAttachmentsTabEnabled = a.getBoolean(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentsTabEnabled,
-                    true
-                )
-
-                val mediaAttachmentsTabIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentsTabIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_media)!!
-
-                val allowAccessToMediaButtonText = a.getText(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToMediaButtonText
-                )?.toString() ?: context.getString(R.string.stream_ui_message_composer_gallery_access)
-
-                val allowAccessToMediaIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToMediaIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_media)!!
-
-                val videoLengthTextVisible = a.getBoolean(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextVisible,
-                    true
-                )
-
-                val videoLengthTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_small)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextColor,
-                        context.getColorCompat(R.color.stream_ui_black)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoLengthTextStyle,
-                        Typeface.NORMAL
-                    )
-                    .build()
-
-                val videoIconVisible = a.getBoolean(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoIconVisible,
-                    true
-                )
-
-                val videoIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_ic_video)!!
-
-                val videoIconDrawableTint =
-                    a.getColorOrNull(R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerVideoIconDrawableTint)
-
-                val mediaAttachmentNoMediaText = a.getString(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaText
-                ) ?: context.getString(R.string.stream_ui_message_composer_no_files)
-
-                val mediaAttachmentNoMediaTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_large)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaTextColor,
-                        context.getColorCompat(R.color.stream_ui_text_color_primary)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerMediaAttachmentNoMediaTextStyle,
-                        Typeface.NORMAL
-                    )
-                    .build()
-
-                /**
-                 * File attachments tab
-                 */
-                val fileAttachmentsTabEnabled = a.getBoolean(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsTabEnabled,
-                    true
-                )
-
-                val fileAttachmentsTabIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsTabIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_file)!!
-
-                val allowAccessToFilesButtonText = a.getText(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToFilesButtonText
-                )?.toString() ?: context.getString(R.string.stream_ui_message_composer_files_access)
-
-                val allowAccessToFilesIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToFilesIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_file)!!
-
-                val recentFilesText = a.getText(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesText
-                )?.toString() ?: context.getString(R.string.stream_ui_message_composer_recent_files)
-
-                val recentFilesTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesTextSize,
-                        context.getDimension(R.dimen.stream_ui_spacing_medium)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesTextColor,
-                        context.getColorCompat(R.color.stream_ui_black)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerRecentFilesTextStyle,
-                        Typeface.BOLD
-                    )
-                    .build()
-
-                val fileManagerIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileManagerIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_ic_file_manager)!!
-
-                val fileAttachmentsNoFilesText = a.getString(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesText
-                ) ?: context.getString(R.string.stream_ui_message_composer_no_files)
-
-                val fileAttachmentsNoFilesTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_large)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesTextColor,
-                        context.getColorCompat(R.color.stream_ui_text_color_primary)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentsNoFilesTextStyle,
-                        Typeface.NORMAL
-                    )
-                    .build()
-
-                val fileAttachmentItemNameTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemNameTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_medium)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemNameTextColor,
-                        context.getColorCompat(R.color.stream_ui_black)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemNameTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemNameTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemNameTextStyle,
-                        Typeface.BOLD
-                    )
-                    .build()
-
-                val fileAttachmentItemSizeTextStyle = TextStyle.Builder(a)
-                    .size(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemSizeTextSize,
-                        context.getDimension(R.dimen.stream_ui_text_small)
-                    )
-                    .color(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemSizeTextColor,
-                        context.getColorCompat(R.color.stream_ui_text_color_secondary)
-                    )
-                    .font(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemSizeTextFontAssets,
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemSizeTextFont
-                    )
-                    .style(
-                        R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemSizeTextStyle,
-                        Typeface.BOLD
-                    )
-                    .build()
-
-                val fileAttachmentItemCheckboxSelectedDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemCheckboxSelectedDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_circle_blue)!!
-
-                val fileAttachmentItemCheckboxDeselectedDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemCheckboxDeselectedDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_ic_file_manager)!!
-
-                val fileAttachmentItemCheckboxTextColor = a.getColor(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerFileAttachmentItemCheckboxTextColor,
-                    context.getColorCompat(R.color.stream_ui_literal_white)
-                )
-
-                /**
-                 * Camera attachments tab
-                 */
-                val cameraAttachmentsTabEnabled = a.getBoolean(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerCameraAttachmentsTabEnabled,
-                    true
-                )
-
-                val cameraAttachmentsTabIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerCameraAttachmentsTabIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_camera)!!
-
-                val allowAccessToCameraButtonText = a.getText(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToCameraButtonText
-                )?.toString() ?: context.getString(R.string.stream_ui_message_composer_camera_access)
-
-                val allowAccessToCameraIconDrawable = a.getDrawable(
-                    R.styleable.AttachmentsPickerDialog_streamUiAttachmentsPickerAllowAccessToCameraIconDrawable
-                ) ?: context.getDrawableCompat(R.drawable.stream_ui_attachment_permission_camera)!!
-
-                return AttachmentsPickerDialogStyle(
-                    attachmentsPickerBackgroundColor = attachmentsPickerBackgroundColor,
-                    allowAccessButtonTextStyle = allowAccessButtonTextStyle,
-                    submitAttachmentsButtonIconDrawable = submitAttachmentsButtonIconDrawable,
-                    attachmentTabToggleButtonStateList = attachmentTabToggleButtonStateList,
-                    // Media attachments tab
-                    mediaAttachmentsTabEnabled = mediaAttachmentsTabEnabled,
-                    mediaAttachmentsTabIconDrawable = mediaAttachmentsTabIconDrawable,
-                    allowAccessToMediaButtonText = allowAccessToMediaButtonText,
-                    allowAccessToMediaIconDrawable = allowAccessToMediaIconDrawable,
-                    videoLengthTextVisible = videoLengthTextVisible,
-                    videoLengthTextStyle = videoLengthTextStyle,
-                    videoIconVisible = videoIconVisible,
-                    videoIconDrawable = videoIconDrawable,
-                    videoIconDrawableTint = videoIconDrawableTint,
-                    mediaAttachmentNoMediaText = mediaAttachmentNoMediaText,
-                    mediaAttachmentNoMediaTextStyle = mediaAttachmentNoMediaTextStyle,
-                    // File attachments tab
-                    fileAttachmentsTabEnabled = fileAttachmentsTabEnabled,
-                    fileAttachmentsTabIconDrawable = fileAttachmentsTabIconDrawable,
-                    allowAccessToFilesButtonText = allowAccessToFilesButtonText,
-                    allowAccessToFilesIconDrawable = allowAccessToFilesIconDrawable,
-                    recentFilesText = recentFilesText,
-                    recentFilesTextStyle = recentFilesTextStyle,
-                    fileManagerIconDrawable = fileManagerIconDrawable,
-                    fileAttachmentsNoFilesText = fileAttachmentsNoFilesText,
-                    fileAttachmentsNoFilesTextStyle = fileAttachmentsNoFilesTextStyle,
-                    fileAttachmentItemNameTextStyle = fileAttachmentItemNameTextStyle,
-                    fileAttachmentItemSizeTextStyle = fileAttachmentItemSizeTextStyle,
-                    fileAttachmentItemCheckboxSelectedDrawable = fileAttachmentItemCheckboxSelectedDrawable,
-                    fileAttachmentItemCheckboxDeselectedDrawable = fileAttachmentItemCheckboxDeselectedDrawable,
-                    fileAttachmentItemCheckboxTextColor = fileAttachmentItemCheckboxTextColor,
-                    // Camera attachments tab
-                    cameraAttachmentsTabEnabled = cameraAttachmentsTabEnabled,
-                    cameraAttachmentsTabIconDrawable = cameraAttachmentsTabIconDrawable,
-                    allowAccessToCameraButtonText = allowAccessToCameraButtonText,
-                    allowAccessToCameraIconDrawable = allowAccessToCameraIconDrawable,
-                ).let(TransformStyle.attachmentsPickerStyleTransformer::transform)
-            }
-        }
-
-        /**
-         * Creates an [AttachmentsPickerDialogStyle] instance with the default values.
-         *
-         * @param context The context to load resources.
-         */
-        public fun createDefault(context: Context): AttachmentsPickerDialogStyle {
-            return invoke(context, null)
-        }
-    }
-}
+)
