--- stream-chat-android-ui-components/src/main/res/values/attrs_attachments_picker_dialog.xml
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-    Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
-
-    Licensed under the Stream License;
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-      https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<resources>
-    <declare-styleable name="AttachmentsPickerDialog">
-        <attr name="streamUiAttachmentsPickerBackgroundColor" format="reference|color" />
-
-        <attr name="streamUiAttachmentsPickerAllowAccessButtonTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerAllowAccessButtonTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerAllowAccessButtonTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerAllowAccessButtonTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerAllowAccessButtonTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerSubmitAttachmentsButtonIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerAttachmentTabToggleButtonStateList" format="color|reference" />
-
-        <!-- Media attachments tab -->
-        <attr name="streamUiAttachmentsPickerMediaAttachmentsTabEnabled" format="boolean" />
-        <attr name="streamUiAttachmentsPickerMediaAttachmentsTabIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToMediaButtonText" format="string" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToMediaIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerVideoLengthTextVisible" format="boolean" />
-
-        <attr name="streamUiAttachmentsPickerVideoLengthTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerVideoLengthTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerVideoLengthTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerVideoLengthTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerVideoLengthTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerVideoIconVisible" format="boolean" />
-        <attr name="streamUiAttachmentsPickerVideoIconDrawable" format="reference" />
-        <!-- The tint of the icon used for video files. -->
-        <attr name="streamUiAttachmentsPickerVideoIconDrawableTint" format="reference|color"/>
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaText" format="string" />
-
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerMediaAttachmentNoMediaTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <!-- File attachments tab -->
-        <attr name="streamUiAttachmentsPickerFileAttachmentsTabEnabled" format="boolean" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsTabIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToFilesButtonText" format="string" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToFilesIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerRecentFilesText" format="string" />
-
-        <attr name="streamUiAttachmentsPickerRecentFilesTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerRecentFilesTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerRecentFilesTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerRecentFilesTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerRecentFilesTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerFileManagerIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesText" format="string" />
-
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentsNoFilesTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemNameTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemNameTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemNameTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemNameTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemNameTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemSizeTextSize" format="dimension" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemSizeTextColor" format="reference|color" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemSizeTextFontAssets" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemSizeTextFont" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemSizeTextStyle" format="enum">
-            <enum name="normal" value="0" />
-            <enum name="bold" value="1" />
-            <enum name="italic" value="2" />
-        </attr>
-
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemCheckboxSelectedDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemCheckboxDeselectedDrawable" format="string" />
-        <attr name="streamUiAttachmentsPickerFileAttachmentItemCheckboxTextColor" format="reference|color" />
-
-        <!-- Camera attachments tab -->
-        <attr name="streamUiAttachmentsPickerCameraAttachmentsTabEnabled" format="boolean" />
-        <attr name="streamUiAttachmentsPickerCameraAttachmentsTabIconDrawable" format="reference" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToCameraButtonText" format="string" />
-        <attr name="streamUiAttachmentsPickerAllowAccessToCameraIconDrawable" format="reference" />
-    </declare-styleable>
-</resources>
